Temporary commit, features appear to be complete. Needs comments and testing.
